backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Embeddings / RAG
sentence-transformers==2.7.0
numpy==1.26.4
tiktoken==0.7.0

# Utility
pydantic==2.6.4
requests==2.32.3
rich==13.7.1

# Optional: JSON cleanup, nicely formatted logs
orjson==3.10.7
